backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.2
python-dotenv>=1.0.1
pydantic>=2.11.0
feedparser>=6.0.11
aiofiles>=24.1.0
